{{ cookiecutter.repo_name }}/requirements.txt
# local package
-e .

# external requirements
awscli
click
coverage
Cython
dvc
jupyterlab
pre-commit
Sphinx
python-dotenv>=0.5.1
optuna

# project-specific stuff
